优化作业支持系统参数和系统函数
